report-5/1

+++ Server_TCP/server.py
@@ -1,7 +1,11 @@
 import os
+import random
 import socket
+import string
 import threading
 import sqlite3
+
+from tqdm import tqdm
 
 IP = "0.0.0.0"
 PORT = 5000
@@ -9,12 +13,12 @@ ADDR = (IP, PORT)
 SIZE = 1024
 FORMAT = "utf-8"
 SERVER_DATA_PATH = "server_data"
+FILE_BLOCK_SIZE = 131072
 
 DB_CONNECTIONS = {}
 DB_CONNECTIONS_LOCK = threading.Lock()
 ACTIVE_SESSIONS = {}
 ACTIVE_SESSIONS_LOCK = threading.Lock()
-
 
 def get_database_connection():
     thread_id = threading.current_thread().ident
@@ -36,32 +40,34 @@ def close_connection(dbconn):
 
 
 def upload_file(conn, data):
-    name, des_path = data[1], data[2]
-    des_path += name.rsplit('/', 1)[0]
+    filename, des_path, file_size = data[1], data[2], int(data[3])
+    des_path += "/" + filename
     filepath = os.path.join(SERVER_DATA_PATH, des_path)
 
-    # send_data = None
-
-    # if os.path.exists(filepath):
-    #     send_data = "2181"
-    #     # conn.send("NO".encode(FORMAT))
-    #     return send_data
-
-    # conn.send("OK".encode(FORMAT))
+    print("To exists")
+    if os.path.exists(filepath):
+        send_data = "2281"
+        conn.send(send_data.encode(FORMAT))
+        return send_data
     
+    conn.send("OK".encode(FORMAT))
+
+    print("To bar")
+    bar = tqdm(
+        range(file_size),
+        f"Receiving {filepath}",
+        unit="B",
+        unit_scale=True,
+        unit_divisor=SIZE,
+    )
+
+    print("To open")
     with open(filepath, "wb") as f:
-        print("Receiving...")
-        # f.write(text)
-        while True:
-            chunk = conn.recv(SIZE)
-            print(chunk)
-            if len(chunk) == 0:
-                print("Len = 0")
-                chunk = chunk.decode(FORMAT)
-            if not chunk:
-                print("Done")
-                break
+        while file_size > 0:
+            chunk = conn.recv(FILE_BLOCK_SIZE)
             f.write(chunk)
+            bar.update(len(chunk))
+            file_size -= len(chunk)
 
     send_data = "1190"
     # conn.send(send_data.encode(FORMAT))
@@ -72,7 +78,7 @@ def download_file(conn, data):
     path = data[1]
     with open(path, "rb") as f:
         while True:
-            chunk = f.read(SIZE)     
+            chunk = f.read(FILE_BLOCK_SIZE)     
             if not chunk:
                 break
             conn.send(chunk)
@@ -178,30 +184,31 @@ def show_my_teams(conn, data, cursor):
 
     if username is not None:
         cursor.execute(
-            "SELECT team_name FROM team_member WHERE account = ?",
+            "SELECT team_code, team_name FROM team WHERE team_name IN (SELECT team_name FROM team_member WHERE account = ?)",
             (username, ),
         )
+        # cursor.execute(
+        #     "SELECT team_name FROM team_member WHERE account = ?",
+        #     (username, ),
+        # )
         result = cursor.fetchall()
 
-        result_code = []
-
-        for row in result:
-            cursor.execute(
-                "SELECT team_code FROM team WHERE team_name = ?",
-                (row[0], ),
-            )
-            code = cursor.fetchone()
-            if code:
-                result_code.append(code)
-        
-        if result:
-            teams = [row[0] for row in result]
-            codes = [row_code[0] for row_code in result_code]
-
-            send_data = "1260\n" + "\n".join(teams).rstrip("\n")
-            send_data += "\n" + "\n".join(codes)
-        else:
+        if not result:
             send_data = "1261\nYou are not a member of any team."
+        else:
+            result_team = []
+            result_code = []
+
+            for row in result:
+                code = row[0]
+                team = row[1]
+                
+                result_code.append(code)
+                result_team.append(team)
+                
+            send_data = "1260\n" + "\n".join(result_team).rstrip("\n")
+            send_data += "\n" + "\n".join(result_code)
+        
     else:
         send_data = "2011"
 
@@ -210,17 +217,17 @@ def show_my_teams(conn, data, cursor):
 
 def show_team_member(conn, data, cursor):
     team_name = data[1]
-    send_data = "1100\n"
+    send_data = "1100"
 
     cursor.execute(
         "SELECT account FROM team_member WHERE team_name = ?",
-        (team_name, ),
+        (team_name, )
     )
 
     result = cursor.fetchall()
 
     for account in result:
-        send_data += "\n".join(account)
+        send_data += "\n" + account[0]
 
     # conn.send(send_data.encode(FORMAT))
     return send_data
@@ -285,6 +292,79 @@ def register(conn, data, cursor, dbconn):
     conn.send("{send_data}\r\n".encode(FORMAT))
     # return send_data
 
+def random_team_code(cursor):
+    characters = string.ascii_letters + string.digits
+    code = "".join(random.choice(characters) for _ in range(6))
+    cursor.execute("Select team_code from team")
+    result = cursor.fetchall()
+    while code in result:
+        characters = string.ascii_letters + string.digits
+        code = "".join(random.choice(characters) for _ in range(6))
+    return code
+
+def create_team(conn, data, cursor, dbconn):
+    print("To create team")
+    team_name, account = data[1], data[2]
+
+    if not team_name:
+        send_data = "2011"
+    else:
+        print("To select team")
+        cursor.execute("SELECT team_name FROM team WHERE team_name = ?", (team_name,))
+        result = cursor.fetchone()
+        if result is not None:
+            send_data = "2041"
+        else:
+            team_path = os.path.join(SERVER_DATA_PATH, team_name)
+            os.makedirs(team_path)
+            team_code = random_team_code(cursor)
+            send_data = "1040\n" + team_code
+            cursor.execute(
+                "INSERT INTO team (leader, team_name, team_code) VALUES (?, ?, ?)",
+                (account, team_name, team_code),
+            )
+            dbconn.commit()
+
+    # conn.send(send_data.encode(FORMAT))
+    print("To return")
+    return send_data
+
+def join_team(conn, data, cursor, dbconn):
+    code = data[1]
+    account = data[2]
+    print("To join team")
+    if code == None or account == None:
+        send_data = "2011"
+    else:
+        cursor.execute(
+            "SELECT team_code, team_name FROM team WHERE team_code = ?", (code,)
+        )
+        result = cursor.fetchone()
+
+        print("To result")
+        if result is None or result[0] is None:
+            send_data = "2061"
+        else:
+            team_name = result[1]
+            cursor.execute(
+                "SELECT account from team_member where team_name = ?", (team_name,)
+            )
+            team_account = cursor.fetchall()
+            print(team_account)
+            print("Account: ", account)
+            if any(account in record for record in team_account):
+                send_data = "2062"
+            else:
+                send_data = "1060"
+                cursor.execute(
+                    "Insert into join_request (sender, team_name) values (?,?)",
+                    (account, team_name),
+                )
+                dbconn.commit()
+    # conn.send(send_data.encode(FORMAT))
+    print("To return")
+    return send_data
+
 ####################################################################
 ####################################################################
 ####################################################################
@@ -298,16 +378,24 @@ def handle_client(conn, addr):
 
     active_session = None
 
-
     while True:
         requests = conn.recv(SIZE).decode(FORMAT)
 
-        if requests == "\r\n":
-            break
-        commands = [request for request in requests.split('\r\n') if request]
-        print(commands)
+        if len(requests) == 0:
+            return
 
-        out = 0
+        if requests == "\r\n":
+            break     
+
+        unknown = None
+        if not requests.endswith("\r\n"):
+            unknown = "3000\r\n"
+            # Sửa
+            commands = [request for request in requests.split('\r\n')[:-1] if request] 
+        else:
+            commands = [request for request in requests.split('\r\n') if request]
+
+        print(commands)
 
         for command in commands:
             data = command.split("\n")
@@ -317,11 +405,11 @@ def handle_client(conn, addr):
                 active_session = login(conn, data, cursor, addr)
             elif cmd == "LOGOUT":
                 active_session = None
-                out = 1
-                break
+                conn.send("1300\r\n".encode(FORMAT))
             elif cmd == "SIGNUP":
                 register(conn, data, cursor, dbconn)
             elif active_session is not None:
+                print("Go to active session")
                 account_id = active_session["account_id"]
                 username = active_session["username"]
                 if cmd == "UPLOAD":
@@ -342,19 +430,24 @@ def handle_client(conn, addr):
                     response = copy_directory(conn, data)
                 elif cmd == "MOVE_FOLDER":
                     response = move_directory(conn, data)
+                elif cmd == "JOIN_TEAM":
+                    response = join_team(conn, data, cursor, dbconn)
+                elif cmd == "CREATE_TEAM":
+                    print("CREATE_TEAM")
+                    response = create_team(conn, data, cursor, dbconn)
                 else:
                     response = "4040"
 
                 response += "\r\n"
+
+                if unknown:
+                    response += unknown
 
                 conn.send(response.encode(FORMAT))
                 print("Response: ", response)
 
             elif len(cmd) == 0:
                 break
-
-        if out == 1:
-            break
 
     print(f"[DISCONNECTED] {addr} disconnected")
     conn.close()
